App/requirements.txt
flask
redis
pymongo
flask_session
python-dotenv
